Applied changes for treating platformSuffix property and refactor.
